+++ CHANGELOG.md
@@ -16,6 +16,7 @@ _This release is scheduled to be released on 2023-07-01._
 - Set Timezone `Europe/Berlin` in unit tests (needed for new formatTime tests)
 - Added no-param-reassign eslint rule and fix warnings
 - updatenotification: Added `sendUpdatesNotifications` feature. Broadcast update with `UPDATES` notification to other modules
+- updatenotification: allow force scanning with `SCAN_UPDATES` notification from other modules
 
 ### Removed